+++ changelog.md
@@ -1,7 +1,10 @@
-7.3.0
+7.4.0
   - added latency and slippage metrics to CopyFactory trade copying API
   - added CopyFactory configuration client method retrieving active resynchronization tasks
   - improved description of CopyFactory account resynchronizing in readme
+  - made it possible to use MetaApi class in interaction tests
+  - renamed tradeCopyingSlippageInPercentPoints -> tradeCopyingSlippageInBasisPoints in CopyFactory history API
+  - added application setting to MetaApi class to make it possible to launch several MetaApi applications in parallel on the same account
 
 7.2.0  
   - Added time fields in broker timezone to objects